pyproject.toml
[tool.poetry]
name = "protonvpn-pmp-qbittorrent"
version = "0.1.0"
description = "A daemon to allocate ProtonVPN port mappings and inform qBittorrent of said mappings."
authors = ["Pierre Bourdon <delroth@gmail.com>", "RibShark <rib@rib.gay>"]
license = "MIT"
readme = "README.md"
packages = [{include = "protonvpn_pmp_qbittorrent"}]

[tool.poetry.scripts]
protonvpn-pmp-qbittorrent = "protonvpn_pmp_qbittorrent.__main__:main"

[tool.poetry.dependencies]
python = "^3.10"
qbittorrent-api = "^2024.5.63"
py-natpmp = "^0.2.5"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
